+++ src/names.c
@@ -572,9 +572,15 @@ r_obj* apply_name_spec(r_obj* name_spec, r_obj* outer, r_obj* inner, r_ssize n) 
     return r_null;
   }
 
-  if (outer == r_null) {
+  if (name_spec_is_inner(name_spec)) {
+    // Ignore `outer` entirely
     return inner;
   }
+  if (outer == r_null) {
+    // `outer` doesn't exist, no need to apply `name_spec`
+    return inner;
+  }
+
   if (r_typeof(outer) != R_TYPE_string) {
     r_stop_internal("`outer` must be a scalar string.");
   }
@@ -647,6 +653,16 @@ r_obj* glue_as_name_spec(r_obj* spec) {
   }
   return vctrs_dispatch1(syms_glue_as_name_spec, fns_glue_as_name_spec,
                          syms_internal_spec, spec);
+}
+
+bool name_spec_is_inner(r_obj* name_spec) {
+  if (!r_is_string(name_spec)) {
+    return false;
+  }
+
+  const char* name_spec_c_string = r_chr_get_c_string(name_spec, 0);
+
+  return !strcmp(name_spec_c_string, "inner");
 }
 
 #define VCTRS_PASTE_BUFFER_MAX_SIZE 4096