+++ tests/testthat/test-slice-chop.R
@@ -938,6 +938,24 @@ test_that("preserves names when inputs are cast to a common type (#1689)", {
   x <- list(c(a = 1), c(b = 2))
   indices <- list(1:2, 3:4)
   expect_named(list_unchop(x, indices = indices, ptype = integer()), c("a", "a", "b", "b"))
+
+  expect_named(
+    list_unchop(
+      list(foo = c(a = 1)),
+      ptype = integer(),
+      name_spec = "inner"
+    ),
+    "a"
+  )
+  expect_named(
+    list_unchop(
+      list(foo = c(a = 1)),
+      ptype = integer(),
+      name_spec = "inner",
+      indices = list(1)
+    ),
+    "a"
+  )
 })
 
 test_that("not all inputs have to be named", {
@@ -1302,24 +1320,65 @@ test_that("list_unchop() falls back for S4 classes with a registered c() method"
   )
 })
 
-test_that("list_unchop() fallback doesn't support `name_spec` or `ptype`", {
+test_that("list_unchop() fallback doesn't support (most) `name_spec` or `ptype`", {
   local_c_foobar()
 
   foo <- structure(foobar(1), foo = "foo")
   bar <- structure(foobar(2), bar = "bar")
 
   expect_snapshot(error = TRUE, {
-    list_unchop(list(foo, bar), name_spec = "{outer}_{inner}")
+    list_unchop(
+      list(foo, bar),
+      indices = list(1, 2),
+      name_spec = "{outer}_{inner}"
+    )
   })
   expect_snapshot(error = TRUE, {
-    list_unchop(list(foo, bar), name_spec = "{outer}_{inner}", error_call = call("foo"))
+    list_unchop(
+      list(foo, bar),
+      indices = list(1, 2),
+      name_spec = "{outer}_{inner}",
+      error_call = call("foo")
+    )
   })
 
   # Used to be an error about `ptype`
   x <- list(foobar(1))
   expect_snapshot(error = TRUE, {
-    list_unchop(x, ptype = "")
+    list_unchop(x, indices = list(1), ptype = "")
   })
+})
+
+test_that("list_unchop() fallback does support `name_spec = 'inner'`", {
+  # Because of how useful it is, and how easy it is to implement!
+  expect_identical(
+    with_c_foobar(list_unchop(
+      list(foobar(1), foobar(2)),
+      indices = list(1, 2),
+      name_spec = "inner"
+    )),
+    foobar_c(c(1, 2))
+  )
+  expect_identical(
+    with_c_foobar(list_unchop(
+      list(x = foobar(1), y = foobar(2)),
+      indices = list(1, 2),
+      name_spec = "inner"
+    )),
+    foobar_c(c(1, 2))
+  )
+  expect_identical(
+    with_c_foobar(list_unchop(
+      list(
+        x = foobar(c(a = 1)),
+        y = foobar(c(b = 2)),
+        z = foobar(3)
+      ),
+      indices = list(1, 2, 3),
+      name_spec = "inner"
+    )),
+    foobar_c(c(a = 1, b = 2, 3))
+  )
 })
 
 test_that("list_unchop() supports numeric S3 indices", {
@@ -1380,6 +1439,26 @@ test_that("can ignore names in `list_unchop()` by providing a `zap()` name-spec 
       ),
       class = "vctrs_error_incompatible_type"
     ))
+  })
+})
+
+test_that("can ignore outer names in `list_unchop()` by providing a 'inner' name-spec (#1988)", {
+  expect_identical(
+    list_unchop(
+      list(x = c(a = 1, 2), y = c(3, b = 4)),
+      indices = list(c(3, 1), c(2, 4)),
+      name_spec = "inner"
+    ),
+    c(2, 3, a = 1, b = 4)
+  )
+
+  # Importantly, outer names are still used in error messages!
+  expect_snapshot(error = TRUE, {
+    list_unchop(
+      list(x = c(a = 1), y = c(b = "2")),
+      indices = list(1, 2),
+      name_spec = "inner"
+    )
   })
 })